Editing title works, added new shortcut CTRL+T for editing title, added confirmation dialogue for deletion

+++ client/src/main/java/client/scenes/NoteOverviewCtrl.java
@@ -8,7 +8,6 @@ import java.util.OptionalLong;
 import java.util.ResourceBundle;
 
 import client.handlers.NoteSearchResult;
-
 import client.services.Markdown;
 import client.utils.NoteUtils;
 import com.google.inject.Inject;
@@ -20,6 +19,7 @@ import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.*;
+import javafx.scene.input.KeyCode;
 import javafx.scene.web.WebView;
 import javafx.stage.Modality;
 import javafx.scene.input.KeyEvent;
@@ -65,7 +65,6 @@ public class NoteOverviewCtrl implements Initializable {
     @FXML
     private TextArea selectedNoteContent;
     private String selectedNoteContentBuffer;
-
     private long selectedNoteId;
 
     @Inject
@@ -118,12 +117,19 @@ public class NoteOverviewCtrl implements Initializable {
         Optional<Note> note = fetchSelectedNote();
         if (note.isEmpty()) return;
         else {
-            server.deleteNote(getSelectedNoteId().getAsLong());
+            String message = "Are you sure you want to delete this note?";
+            String title = "Confirm deletion";
+            int choice = JOptionPane.showConfirmDialog(null, message, title, JOptionPane.YES_NO_OPTION,JOptionPane.QUESTION_MESSAGE);
+            if (choice == JOptionPane.YES_OPTION) {
+                server.deleteNote(getSelectedNoteId().getAsLong());
+                selectedNoteTitle.setText(" ");
+                selectedNoteContent.setText(" ");
+                refresh();
+                selectedNoteContent.setDisable(true);
+            } else {
+                refresh();
+            }
         }
-        selectedNoteTitle.setText(" ");
-        selectedNoteContent.setText(" ");
-        refresh();
-        selectedNoteContent.setDisable(true);
     }
 
     public void emptySearchText() {
@@ -289,6 +295,7 @@ public class NoteOverviewCtrl implements Initializable {
     /**
      * Currently only has a keyboard shortcut for refreshing/searching
      * more shortcuts can be added in the future.
+     *
      * @param e
      */
     public void keyPressed(KeyEvent e) {
@@ -303,12 +310,23 @@ public class NoteOverviewCtrl implements Initializable {
                 addNote();
                 break;
             default:
-                break;
+                if (e.getCode() == KeyCode.T && e.isControlDown()) {
+                    setTitle();
+                    break;
+                } else {
+                    break;
+                }
         }
     }
 
-    public void title() {
-        mainCtrl.getNewCtrl().newTitle(table.getSelectionModel().getSelectedItem());
+    /**
+     * Sets the title of the selected note.
+     */
+    public void setTitle() {
+        Note note = table.getSelectionModel().getSelectedItem();
+        if (note != null) {
+            mainCtrl.getNewCtrl().newTitle(table.getSelectionModel().getSelectedItem());
+        }
     }
 
     public void empty() {
